Changed mov from ints to doubles, started vectors/nodes

+++ src/juego/Game.java
@@ -139,9 +139,9 @@ public class Game extends Canvas implements Runnable {
         }
         
         screen.clear();
-        int xScroll = player.getX() - screen.width / 2;
-        int yScroll = player.getY() - screen.height / 2;
-        level.render(xScroll, yScroll, screen);
+        double xScroll = player.getX() - screen.width / 2;
+        double yScroll = player.getY() - screen.height / 2;
+        level.render((int)xScroll, (int)yScroll, screen);
 
         // Copy the pixels we have in Screen.java to the pixels array here/*
         for (int i = 0; i < pixels.length; i++) {
